ApiBlockTest: delete a block only if it exist
Fix cruise control, follow up r89071

+++ tests/phpunit/includes/api/ApiBlockTest.php
@@ -30,7 +30,9 @@ class ApiBlockTest extends ApiTestSetup {
 
 	function removeDBData() {
 		$block = Block::newFromTarget('UTBlockee');
-		$block->delete();	
+		if( $block ) {
+			$block->delete();
+		}
 	}
